rename the wechaty_plugin_contrib
