--- src/python_wechaty_plugin_contrib/__init__.py
@@ -1,11 +1,0 @@
-"""import all plugins"""
-from python_wechaty_plugin_contrib.daily_plugin import DailyPlugin
-from python_wechaty_plugin_contrib.ding_dong_plugin import DingDongPlugin
-from python_wechaty_plugin_contrib.weather_plugin import WeatherPlugin
-
-
-__all__ = [
-    'DailyPlugin',
-    'DingDongPlugin',
-    'WeatherPlugin'
-]
