+++ src/wechaty_plugin_contrib/__init__.py
@@ -1,0 +1,11 @@
+"""import all plugins"""
+from wechaty_plugin_contrib.daily_plugin import DailyPlugin
+from wechaty_plugin_contrib.ding_dong_plugin import DingDongPlugin
+from wechaty_plugin_contrib.weather_plugin import WeatherPlugin
+
+
+__all__ = [
+    'DailyPlugin',
+    'DingDongPlugin',
+    'WeatherPlugin'
+]
