+++ src/wechaty_plugin_contrib/daily_plugin.py
