+++ src/wechaty_plugin_contrib/ding_dong_plugin.py
