+++ src/wechaty_plugin_contrib/ding_dong_rank_plugin.py
