+++ src/wechaty_plugin_contrib/weather_plugin.py
